+++ src/apps/ScopeChangeRequest/api/ScopeChange/getPhases.ts
@@ -1,6 +1,6 @@
 import { httpClient } from '../../../../Core/Client/Functions';
 import { throwOnError } from '../../functions/throwError';
-import { QueryContext } from '../../keys/queries';
+import { QueryContext } from '../../types/query/queryContext';
 
 interface Phase {
     name: string;
