+++ src/apps/ScopeChangeRequest/api/ScopeChange/getScopes.ts
@@ -1,7 +1,6 @@
 import { httpClient } from '@equinor/lighthouse-portal-client';
 import { throwOnError } from '../../functions/throwError';
-import { QueryContext } from '../../keys/queries';
-
+import { QueryContext } from '../../types/query/queryContext';
 
 interface Scope {
     name: string;
