+++ src/apps/ScopeChangeRequest/types/query/queryContext.ts
@@ -1,0 +1,3 @@
+export interface QueryContext {
+    signal?: AbortSignal;
+}
